backend/requirements.txt
Django>=4.0
djangorestframework>=3.12
openai>=0.27.0
python-dotenv>=0.19.2